backend/requirements.txt
werkzeug==3.0.6
flask==3.0.2
firebase_admin<6.0
pillow==10.2.0
opencv-python==4.6.0.66
numpy==1.24.4
matplotlib==3.7.4
opencv-contrib-python==4.6.0.66
tensorflow==2.10.0
protobuf<4
gunicorn>=20.0.0,<21.0.0
google-cloud-firestore==2.3.0